+++ yellowstone-grpc-geyser/src/grpc.rs
@@ -728,7 +728,7 @@ pub struct GrpcService {
     config_filters: Arc<ConfigGrpcFilters>,
     blocks_meta: Option<BlockMetaStorage>,
     subscribe_id: AtomicUsize,
-    snapshot_rx: Mutex<Option<crossbeam_channel::Receiver<Option<Message>>>>,
+    snapshot_rx: Mutex<Option<crossbeam_channel::Receiver<Option<Box<Message>>>>>,
     broadcast_tx: broadcast::Sender<(CommitmentLevel, Arc<Vec<Arc<Message>>>)>,
     debug_clients_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
 }
@@ -741,7 +741,7 @@ impl GrpcService {
         debug_clients_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
         is_reload: bool,
     ) -> anyhow::Result<(
-        Option<crossbeam_channel::Sender<Option<Message>>>,
+        Option<crossbeam_channel::Sender<Option<Box<Message>>>>,
         mpsc::UnboundedSender<Arc<Message>>,
         Arc<Notify>,
     )> {
@@ -1125,7 +1125,7 @@ impl GrpcService {
         config_filters: Arc<ConfigGrpcFilters>,
         stream_tx: mpsc::Sender<TonicResult<SubscribeUpdate>>,
         mut client_rx: mpsc::UnboundedReceiver<Option<Filter>>,
-        mut snapshot_rx: Option<crossbeam_channel::Receiver<Option<Message>>>,
+        mut snapshot_rx: Option<crossbeam_channel::Receiver<Option<Box<Message>>>>,
         mut messages_rx: broadcast::Receiver<(CommitmentLevel, Arc<Vec<Arc<Message>>>)>,
         debug_client_tx: Option<mpsc::UnboundedSender<DebugClientMessage>>,
         drop_client: impl FnOnce(),
@@ -1256,7 +1256,7 @@ impl GrpcService {
         endpoint: &str,
         stream_tx: &mpsc::Sender<TonicResult<SubscribeUpdate>>,
         client_rx: &mut mpsc::UnboundedReceiver<Option<Filter>>,
-        snapshot_rx: crossbeam_channel::Receiver<Option<Message>>,
+        snapshot_rx: crossbeam_channel::Receiver<Option<Box<Message>>>,
         is_alive: &mut bool,
         filter: &mut Filter,
     ) {